requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.3
djangorestframework==3.14.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
Pygments==2.3.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
pytz==2018.9
PyYAML==5.4.1
redgreenunittest==0.1.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
